CC-1909: Refactor, move ImageMap into its own file, and move Options stuff from api.go into options.go
